refactor: BoardConverter 패키지 이동
